+++ package.json
@@ -77,6 +77,7 @@
     "@babel/preset-env": "^7.12.7",
     "@babel/preset-react": "^7.12.7",
     "@babel/preset-typescript": "^7.12.7",
+    "@ladle/react": "^0.8.3",
     "@rollup/plugin-babel": "^5.2.1",
     "@rollup/plugin-image": "^2.1.1",
     "@rollup/plugin-json": "^4.1.0",
@@ -191,7 +192,8 @@
     "test": "jest",
     "types": "tsc --strict",
     "validate-translations": "node scripts/validate-translations.js",
-    "semantic-release": "semantic-release"
+    "semantic-release": "semantic-release",
+    "browse-examples": "export $(cat .env | sed 's/^/VITE_/g' | sed 's/#.*//g' | xargs) && ladle serve"
   },
   "resolutions": {
     "ast-types": "^0.14.0"
