+++ src/components/Reactions/utils/utils.tsx
@@ -1,4 +1,4 @@
-export const isMutableRef = <T>(
+export const isMutableRef = <T,>(
   ref: React.ForwardedRef<T> | null,
 ): ref is React.MutableRefObject<T> => {
   if (ref) {
